requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai==1.3.0
icalendar==5.0.11
pydantic==2.4.2
python-dotenv==1.0.0
